cleanup and tested working
